Close #75: Implement sold data of products.

+++ modules/EasyAbp.EShop.Products/src/EasyAbp.EShop.Products.Application.Contracts/EasyAbp/EShop/Products/ProductInventories/Dtos/ProductInventoryDto.cs
@@ -12,6 +12,6 @@ namespace EasyAbp.EShop.Products.ProductInventories.Dtos
 
         public int Inventory { get; set; }
 
-        public int Sold { get; set; }
+        public long Sold { get; set; }
     }
 }
